removed nodemon and moved all legacy stuff to legacy
